move some common methods into the abstract test runner
